refactor(message): message enum들을 message 패키지로 이동
