+++ src/main/java/christmas/message/Error.java
@@ -1,6 +1,6 @@
-package christmas.view;
+package christmas.message;
 
-public enum ErrorMessage {
+public enum Error {
 
     INVALUD_DATE("유효하지 않은 날짜입니다."),
     INVALUD_ORDER("유효하지 않은 주문입니다."),
@@ -12,7 +12,7 @@ public enum ErrorMessage {
     private final static String POSTFIX = " 다시 입력해 주세요.";
     private final String message;
 
-    ErrorMessage(String message) {
+    Error(String message) {
         this.message = message;
     }
